All surfaces apart from card now adhere to an interface that derives from or is the INgxFluentDesignSurface interface.
Moved Orchestrator, animations and some other things into the surface module as it's the only place they get used.

You can now pass a class list to the orchestrator and it'll apply these to the body of the document. If we need other behaviour then it might be best to create an interface from this orchestrator and implement it separately, but for now all surfaces only have to stop body scrolling when they're open.